Moved the countries list file in the config/buffalogs folder (#385)
